refactor: Separate User and Meeting services from service implementations
